~move SerialDescriptorBuilder to a separate package to preserve file history
