+++ runtime/commonMain/src/kotlinx/serialization/builtins/LongAsStringSerializer.kt
@@ -5,6 +5,7 @@
 package kotlinx.serialization.builtins
 
 import kotlinx.serialization.*
+import kotlinx.serialization.descriptors.*
 
 
 /**
