+++ runtime/commonMain/src/kotlinx/serialization/internal/Enums.kt
@@ -5,6 +5,7 @@
 package kotlinx.serialization.internal
 
 import kotlinx.serialization.*
+import kotlinx.serialization.descriptors.*
 
 /*
  * Descriptor used for explicitly serializable enums by the plugin.
