+++ runtime/commonMain/src/kotlinx/serialization/internal/ObjectSerializer.kt
@@ -5,6 +5,7 @@
 package kotlinx.serialization.internal
 
 import kotlinx.serialization.*
+import kotlinx.serialization.descriptors.*
 
 /**
  * Serializer for Kotlin's singletons (denoted by `object` keyword).
