+++ runtime/commonMain/src/kotlinx/serialization/internal/Tuples.kt
@@ -6,6 +6,7 @@
 package kotlinx.serialization.internal
 
 import kotlinx.serialization.*
+import kotlinx.serialization.descriptors.*
 import kotlin.native.concurrent.*
 
 @SharedImmutable
