+++ runtime/commonMain/src/kotlinx/serialization/json/JsonElementSerializer.kt
@@ -6,6 +6,7 @@ package kotlinx.serialization.json
 
 import kotlinx.serialization.*
 import kotlinx.serialization.builtins.*
+import kotlinx.serialization.descriptors.*
 import kotlinx.serialization.internal.*
 
 /**
