+++ runtime/commonMain/src/kotlinx/serialization/json/JsonTransformingSerializer.kt
@@ -5,6 +5,7 @@
 package kotlinx.serialization.json
 
 import kotlinx.serialization.*
+import kotlinx.serialization.descriptors.*
 import kotlinx.serialization.json.internal.*
 
 /**
