+++ runtime/commonTest/src/kotlinx/serialization/EnumSerializationTest.kt
@@ -4,6 +4,7 @@
 
 package kotlinx.serialization
 
+import kotlinx.serialization.descriptors.*
 import kotlinx.serialization.json.JsonTestBase
 import kotlinx.serialization.test.*
 import kotlin.test.*
