+++ runtime/commonTest/src/kotlinx/serialization/SerialDescriptorBuilderTest.kt
@@ -5,6 +5,7 @@
 package kotlinx.serialization
 
 import kotlinx.serialization.builtins.*
+import kotlinx.serialization.descriptors.*
 import kotlinx.serialization.test.*
 import kotlin.test.*
