+++ runtime/commonTest/src/kotlinx/serialization/SerialDescriptorEqualityTest.kt
@@ -5,6 +5,7 @@
 package kotlinx.serialization
 
 import kotlinx.serialization.builtins.*
+import kotlinx.serialization.descriptors.*
 import kotlin.test.*
 
 class SerialDescriptorEqualityTest {
