+++ runtime/commonTest/src/kotlinx/serialization/SerialDescriptorSpecificationTest.kt
@@ -6,6 +6,7 @@ package kotlinx.serialization
 
 import kotlinx.serialization.CompositeDecoder.Companion.UNKNOWN_NAME
 import kotlinx.serialization.builtins.*
+import kotlinx.serialization.descriptors.*
 import kotlin.test.*
 
 class SerialDescriptorSpecificationTest {
