+++ runtime/commonTest/src/kotlinx/serialization/features/GenericCustomSerializerTest.kt
@@ -6,6 +6,7 @@ package kotlinx.serialization.features
 
 import kotlinx.serialization.*
 import kotlinx.serialization.builtins.*
+import kotlinx.serialization.descriptors.*
 import kotlinx.serialization.internal.*
 import kotlinx.serialization.json.*
 import kotlin.test.*
