+++ runtime/commonTest/src/kotlinx/serialization/json/JsonCustomSerializersTest.kt
@@ -7,6 +7,7 @@ package kotlinx.serialization.json
 
 import kotlinx.serialization.*
 import kotlinx.serialization.builtins.*
+import kotlinx.serialization.descriptors.*
 import kotlinx.serialization.modules.*
 import kotlin.test.*
