+++ runtime/commonTest/src/kotlinx/serialization/json/JsonEncoderDecoderRecursiveTest.kt
@@ -5,6 +5,7 @@
 package kotlinx.serialization.json
 
 import kotlinx.serialization.*
+import kotlinx.serialization.descriptors.*
 import kotlin.test.*
 
 class JsonEncoderDecoderRecursiveTest : JsonTestBase() {
