+++ runtime/commonTest/src/kotlinx/serialization/json/MapLikeSerializerTest.kt
@@ -6,6 +6,7 @@ package kotlinx.serialization.json
 
 import kotlinx.serialization.*
 import kotlinx.serialization.builtins.*
+import kotlinx.serialization.descriptors.*
 import kotlin.test.*
 
 class MapLikeSerializerTest : JsonTestBase() {
