+++ runtime/commonTest/src/kotlinx/serialization/modules/ModuleBuildersTest.kt
@@ -8,6 +8,7 @@ import kotlinx.serialization.*
 import kotlinx.serialization.PolyBase
 import kotlinx.serialization.PolyDerived
 import kotlinx.serialization.builtins.*
+import kotlinx.serialization.descriptors.*
 import kotlinx.serialization.test.*
 import kotlin.reflect.*
 import kotlin.test.*
