+++ runtime/jvmTest/src/kotlinx/serialization/features/SerializerByTypeTest.kt
@@ -6,6 +6,7 @@ package kotlinx.serialization.features
 
 import kotlinx.serialization.*
 import kotlinx.serialization.builtins.*
+import kotlinx.serialization.descriptors.*
 import kotlinx.serialization.json.*
 import org.junit.Test
 import java.lang.reflect.*
